+++ hosts/jak-hl/packages-fonts.nix
@@ -1,18 +1,22 @@
 # 💫 https://github.com/LinuxBeginnings 💫 #
 # Packages for this host only
-{pkgs, ...}: let
+{ inputs, pkgs, ... }:
+let
   python-packages = pkgs.python3.withPackages (
-    ps:
-      with ps; [
-        requests
-        pyquery # needed for hyprland-dots Weather script
-      ]
+    ps: with ps; [
+      requests
+      pyquery # needed for hyprland-dots Weather script
+    ]
   );
-in {
+in
+{
   nixpkgs.config.allowUnfree = true;
 
   environment.systemPackages =
     (with pkgs; [
+
+      inputs.antigravity-cli-repo.packages.${pkgs.stdenv.hostPlatform.system}.antigravity-cli
+
       # System Packages
       google-chrome
     ])