+++ hosts/nixos/packages-fonts.nix
@@ -1,14 +1,15 @@
 # 💫 https://github.com/LinuxBeginnings 💫 #
 # Packages and Fonts config including the "programs" options
-{pkgs, ...}: let
+{ pkgs, ... }:
+let
   python-packages = pkgs.python3.withPackages (
-    ps:
-      with ps; [
-        requests
-        pyquery # needed for hyprland-dots Weather script
-      ]
+    ps: with ps; [
+      requests
+      pyquery # needed for hyprland-dots Weather script
+    ]
   );
-in {
+in
+{
   nixpkgs.config.allowUnfree = true;
 
   environment.systemPackages =