+++ modules/home/cli/git.nix
@@ -1,4 +1,5 @@
-{lib, ...}: {
+{ lib, ... }:
+{
   programs.git = {
     enable = true;
     signing.format = null;