+++ modules/home/default.nix
@@ -1,4 +1,5 @@
-{...}: {
+{ ... }:
+{
   imports = [
     ./terminals/tmux.nix
     #./terminals/ghostty.nix