+++ modules/local-hardware-clock.nix
@@ -11,12 +11,14 @@
   config,
   ...
 }:
-with lib; let
+with lib;
+let
   cfg = config.local.hardware-clock;
-in {
+in
+{
   options.local.hardware-clock = {
     enable = mkEnableOption "Change Hardware Clock To Local Time";
   };
 
-  config = mkIf cfg.enable {time.hardwareClockInLocalTime = true;};
+  config = mkIf cfg.enable { time.hardwareClockInLocalTime = true; };
 }