+++ modules/vm-guest-services.nix
@@ -11,9 +11,11 @@
   config,
   ...
 }:
-with lib; let
+with lib;
+let
   cfg = config.vm.guest-services;
-in {
+in
+{
   options.vm.guest-services = {
     enable = mkEnableOption "Enable Virtual Machine Guest Services";
   };